+++ ServerINes/Services/Features/Locations/Commands/DeleteLocation/DeleteLocationHandler.cs
@@ -1,4 +1,5 @@
-﻿using INest.Services.Tracker;
+﻿using INest.Exceptions;
+using INest.Services.Tracker;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using static INest.Constants.LocalizationConstants;
@@ -19,15 +20,25 @@ namespace INest.Services.Features.Locations.Commands.DeleteLocation
         public async Task<bool> Handle(DeleteLocationCommand request, CancellationToken cancellationToken)
         {
             var location = await _context.StorageLocations
-                .FirstOrDefaultAsync(l => l.Id == request.LocationId && l.UserId == request.UserId, cancellationToken);
+                .Include(l => l.Children)
+                .Include(l => l.Items)
+                .FirstOrDefaultAsync(l => l.Id == request.Id && l.UserId == request.UserId, cancellationToken);
 
             if (location == null)
-                throw new KeyNotFoundException(LOCATIONS.ERRORS.NOT_FOUND);
+            {
+                throw new AppException(LOCATIONS.ERRORS.NOT_FOUND, 404);
+            }
+
+            if (location.Children.Any() || location.Items.Any())
+            {
+                throw new AppException(ERRORS.NOT_EMPTY, 400);
+            }
 
             _context.StorageLocations.Remove(location);
             await _context.SaveChangesAsync(cancellationToken);
 
             _tracker.InvalidateUserCache(request.UserId);
+
             return true;
         }
     }
